[Test] Remove macros in test in favor of fns (#552)
