+++ core/Cargo.toml
@@ -44,7 +44,6 @@ bitcoin-script = { workspace = true }
 
 [features]
 default = []
-testing = []
 
 [[bin]]
 name = "server"
