+++ core/src/actor.rs
@@ -481,13 +481,10 @@ mod tests {
     use crate::builder::transaction::input::SpendableTxIn;
     use crate::builder::transaction::output::UnspentTxOut;
     use crate::builder::transaction::{TransactionType, TxHandler, TxHandlerBuilder};
-    use crate::config::BridgeConfig;
+
     use crate::rpc::clementine::NormalSignatureKind;
-    use crate::utils::{initialize_logger, SECP};
-    use crate::{
-        actor::WinternitzDerivationPath, create_test_config_with_thread_name, database::Database,
-        initialize_database,
-    };
+    use crate::utils::SECP;
+    use crate::{actor::WinternitzDerivationPath, test::common::*};
     use bitcoin::secp256k1::{schnorr, Message, SecretKey};
 
     use bitcoin::sighash::TapSighashType;
@@ -738,7 +735,7 @@ mod tests {
 
     #[tokio::test]
     async fn derive_winternitz_pk_uniqueness() {
-        let config = create_test_config_with_thread_name!(None);
+        let config = create_test_config_with_thread_name(None).await;
         let actor = Actor::new(
             config.secret_key,
             config.winternitz_secret_key,
@@ -757,7 +754,7 @@ mod tests {
 
     #[tokio::test]
     async fn derive_winternitz_pk_fixed_pk() {
-        let config = create_test_config_with_thread_name!(None);
+        let config = create_test_config_with_thread_name(None).await;
         let actor = Actor::new(
             config.secret_key,
             Some(
@@ -814,7 +811,7 @@ mod tests {
 
     #[tokio::test]
     async fn sign_winternitz_signature() {
-        let config = create_test_config_with_thread_name!(None);
+        let config = create_test_config_with_thread_name(None).await;
         let actor = Actor::new(
             config.secret_key,
             Some(
