+++ core/src/bitcoin_syncer.rs
@@ -339,11 +339,8 @@ pub async fn start_bitcoin_syncer(
 #[cfg(test)]
 mod tests {
     use crate::builder::transaction::DEFAULT_SEQUENCE;
-    use crate::extended_rpc::ExtendedRpc;
-    use crate::{
-        config::BridgeConfig, database::Database, initialize_database, utils::initialize_logger,
-    };
-    use crate::{create_regtest_rpc, create_test_config_with_thread_name};
+
+    use crate::{database::Database, test::common::*};
     use bitcoin::absolute::Height;
     use bitcoin::hashes::Hash;
     use bitcoin::transaction::Version;
@@ -354,8 +351,8 @@ mod tests {
     #[tokio::test]
     #[serial_test::serial]
     async fn get_block_info_from_height() {
-        let mut config = create_test_config_with_thread_name!(None);
-        let regtest = create_regtest_rpc!(config);
+        let mut config = create_test_config_with_thread_name(None).await;
+        let regtest = create_regtest_rpc(&mut config).await;
         let rpc = regtest.rpc().clone();
 
         rpc.mine_blocks(1).await.unwrap();
@@ -384,9 +381,9 @@ mod tests {
     #[tokio::test]
     #[serial_test::serial]
     async fn save_get_transaction_spent_utxos() {
-        let mut config = create_test_config_with_thread_name!(None);
+        let mut config = create_test_config_with_thread_name(None).await;
         let db = Database::new(&config).await.unwrap();
-        let regtest = create_regtest_rpc!(config);
+        let regtest = create_regtest_rpc(&mut config).await;
         let rpc = regtest.rpc().clone();
 
         let mut dbtx = db.begin_transaction().await.unwrap();
@@ -443,9 +440,9 @@ mod tests {
     #[tokio::test]
     #[serial_test::serial]
     async fn save_get_block() {
-        let mut config = create_test_config_with_thread_name!(None);
+        let mut config = create_test_config_with_thread_name(None).await;
         let db = Database::new(&config).await.unwrap();
-        let regtest = create_regtest_rpc!(config);
+        let regtest = create_regtest_rpc(&mut config).await;
         let rpc = regtest.rpc().clone();
 
         let mut dbtx = db.begin_transaction().await.unwrap();
@@ -477,9 +474,9 @@ mod tests {
     #[tokio::test]
     #[serial_test::serial]
     async fn set_initial_block_info_if_not_exists() {
-        let mut config = create_test_config_with_thread_name!(None);
+        let mut config = create_test_config_with_thread_name(None).await;
         let db = Database::new(&config).await.unwrap();
-        let regtest = create_regtest_rpc!(config);
+        let regtest = create_regtest_rpc(&mut config).await;
         let rpc = regtest.rpc().clone();
 
         let mut dbtx = db.begin_transaction().await.unwrap();
@@ -513,9 +510,9 @@ mod tests {
     #[tokio::test]
     #[serial_test::serial]
     async fn fetch_new_blocks_forward() {
-        let mut config = create_test_config_with_thread_name!(None);
+        let mut config = create_test_config_with_thread_name(None).await;
         let db = Database::new(&config).await.unwrap();
-        let regtest = create_regtest_rpc!(config);
+        let regtest = create_regtest_rpc(&mut config).await;
         let rpc = regtest.rpc().clone();
 
         let mut dbtx = db.begin_transaction().await.unwrap();
@@ -549,9 +546,9 @@ mod tests {
     #[tokio::test]
     #[serial_test::serial]
     async fn fetch_new_blocks_backwards() {
-        let mut config = create_test_config_with_thread_name!(None);
+        let mut config = create_test_config_with_thread_name(None).await;
         let db = Database::new(&config).await.unwrap();
-        let regtest = create_regtest_rpc!(config);
+        let regtest = create_regtest_rpc(&mut config).await;
         let rpc = regtest.rpc().clone();
 
         // Prepare chain.
@@ -598,9 +595,9 @@ mod tests {
     #[tokio::test]
     #[serial_test::serial]
     async fn set_non_canonical_block_hashes() {
-        let mut config = create_test_config_with_thread_name!(None);
+        let mut config = create_test_config_with_thread_name(None).await;
         let db = Database::new(&config).await.unwrap();
-        let regtest = create_regtest_rpc!(config);
+        let regtest = create_regtest_rpc(&mut config).await;
         let rpc = regtest.rpc().clone();
 
         let hashes = rpc.mine_blocks(4).await.unwrap();
@@ -644,9 +641,9 @@ mod tests {
     #[tokio::test]
     #[serial_test::serial]
     async fn start_bitcoin_syncer_new_block_mined() {
-        let mut config = create_test_config_with_thread_name!(None);
+        let mut config = create_test_config_with_thread_name(None).await;
         let db = Database::new(&config).await.unwrap();
-        let regtest = create_regtest_rpc!(config);
+        let regtest = create_regtest_rpc(&mut config).await;
         let rpc = regtest.rpc().clone();
 
         rpc.mine_blocks(1).await.unwrap();
