+++ core/src/builder/script.rs
@@ -359,7 +359,7 @@ fn get_script_from_arr<T: SpendableScript>(
 mod tests {
     use crate::actor::{Actor, WinternitzDerivationPath};
     use crate::extended_rpc::ExtendedRpc;
-    use crate::{create_regtest_rpc, create_test_config_with_thread_name, utils};
+    use crate::utils;
     use std::sync::Arc;
 
     use super::*;
@@ -566,15 +566,13 @@ mod tests {
         (builder, address)
     }
 
-    use crate::{
-        config::BridgeConfig, database::Database, initialize_database, utils::initialize_logger,
-    };
+    use crate::test::common::*;
 
     #[tokio::test]
 
     async fn test_checksig_spendable() {
-        let mut config = create_test_config_with_thread_name!(None);
-        let regtest = create_regtest_rpc!(config);
+        let mut config = create_test_config_with_thread_name(None).await;
+        let regtest = create_regtest_rpc(&mut config).await;
         let rpc = regtest.rpc().clone();
 
         let kp = bitcoin::secp256k1::Keypair::new(&SECP, &mut rand::thread_rng());
@@ -612,8 +610,8 @@ mod tests {
 
     #[tokio::test]
     async fn test_winternitz_commit_spendable() {
-        let mut config = create_test_config_with_thread_name!(None);
-        let regtest = create_regtest_rpc!(config);
+        let mut config = create_test_config_with_thread_name(None).await;
+        let regtest = create_regtest_rpc(&mut config).await;
         let rpc = regtest.rpc();
 
         let kp = bitcoin::secp256k1::Keypair::new(&SECP, &mut rand::thread_rng());
@@ -662,8 +660,8 @@ mod tests {
 
     #[tokio::test]
     async fn test_timelock_script_spendable() {
-        let mut config = create_test_config_with_thread_name!(None);
-        let regtest = create_regtest_rpc!(config);
+        let mut config = create_test_config_with_thread_name(None).await;
+        let regtest = create_regtest_rpc(&mut config).await;
         let rpc = regtest.rpc();
 
         let kp = bitcoin::secp256k1::Keypair::new(&SECP, &mut rand::thread_rng());
@@ -707,8 +705,8 @@ mod tests {
     #[tokio::test]
 
     async fn test_preimage_reveal_script_spendable() {
-        let mut config = create_test_config_with_thread_name!(None);
-        let regtest = create_regtest_rpc!(config);
+        let mut config = create_test_config_with_thread_name(None).await;
+        let regtest = create_regtest_rpc(&mut config).await;
         let rpc = regtest.rpc().clone();
         let kp = bitcoin::secp256k1::Keypair::new(&SECP, &mut rand::thread_rng());
         let xonly_pk = kp.public_key().x_only_public_key().0;
@@ -750,8 +748,8 @@ mod tests {
     #[tokio::test]
 
     async fn test_deposit_script_spendable() {
-        let mut config = create_test_config_with_thread_name!(None);
-        let regtest = create_regtest_rpc!(config);
+        let mut config = create_test_config_with_thread_name(None).await;
+        let regtest = create_regtest_rpc(&mut config).await;
         let rpc = regtest.rpc().clone();
 
         let kp = bitcoin::secp256k1::Keypair::new(&SECP, &mut rand::thread_rng());
