+++ core/src/database/operator.rs
@@ -588,14 +588,14 @@ impl Database {
     ) -> Result<(), BridgeError> {
         let query = sqlx::query(
             "WITH deposit AS (
-            INSERT INTO deposits (deposit_outpoint) 
-            VALUES ($1) 
-            ON CONFLICT DO NOTHING 
+            INSERT INTO deposits (deposit_outpoint)
+            VALUES ($1)
+            ON CONFLICT DO NOTHING
             RETURNING deposit_id
             )
             INSERT INTO deposit_signatures (deposit_id, operator_idx, sequential_collateral_idx, kickoff_idx, signatures)
             VALUES (
-            (SELECT deposit_id FROM deposit UNION SELECT deposit_id FROM deposits WHERE deposit_outpoint = $1), 
+            (SELECT deposit_id FROM deposit UNION SELECT deposit_id FROM deposits WHERE deposit_outpoint = $1),
             $2, $3, $4, $5
             );"
         )
@@ -767,20 +767,17 @@ mod tests {
     use bitcoin::secp256k1::schnorr;
     use bitcoin::{Amount, OutPoint, ScriptBuf, TxOut, Txid};
 
-    use crate::extended_rpc::ExtendedRpc;
     use crate::operator::Operator;
     use crate::rpc::clementine::{
         DepositSignatures, NormalSignatureKind, TaggedSignature, WatchtowerSignatureKind,
     };
-    use crate::{
-        config::BridgeConfig, database::Database, initialize_database, utils::initialize_logger,
-    };
-    use crate::{create_regtest_rpc, create_test_config_with_thread_name, UTXO};
+    use crate::UTXO;
+    use crate::{database::Database, test::common::*};
     use std::str::FromStr;
 
     #[tokio::test]
     async fn save_get_operators() {
-        let config = create_test_config_with_thread_name!(None);
+        let config = create_test_config_with_thread_name(None).await;
         let database = Database::new(&config).await.unwrap();
         let mut ops = Vec::new();
         for i in 0..2 {
@@ -825,7 +822,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_save_get_public_hashes() {
-        let config = create_test_config_with_thread_name!(None);
+        let config = create_test_config_with_thread_name(None).await;
         let database = Database::new(&config).await.unwrap();
 
         let operator_idx = 0;
@@ -865,7 +862,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_database_gets_previously_saved_operator_take_signature() {
-        let config = create_test_config_with_thread_name!(None);
+        let config = create_test_config_with_thread_name(None).await;
         let database = Database::new(&config).await.unwrap();
 
         let deposit_outpoint = OutPoint::null();
@@ -903,7 +900,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_deposit_kickoff_generator_tx_0() {
-        let config = create_test_config_with_thread_name!(None);
+        let config = create_test_config_with_thread_name(None).await;
         let db = Database::new(&config).await.unwrap();
 
         let raw_hex = "02000000000101eb87b1a80d47b7f5bd5082b77653f5ca37e566951742b80c361875ba0e5c478f0a00000000fdffffff0ca086010000000000225120b23da6d2e0390018b953f7d74e3582da4da30fd0fd157cc84a2d2753003d1ca3a086010000000000225120b23da6d2e0390018b953f7d74e3582da4da30fd0fd157cc84a2d2753003d1ca3a086010000000000225120b23da6d2e0390018b953f7d74e3582da4da30fd0fd157cc84a2d2753003d1ca3a086010000000000225120b23da6d2e0390018b953f7d74e3582da4da30fd0fd157cc84a2d2753003d1ca3a086010000000000225120b23da6d2e0390018b953f7d74e3582da4da30fd0fd157cc84a2d2753003d1ca3a086010000000000225120b23da6d2e0390018b953f7d74e3582da4da30fd0fd157cc84a2d2753003d1ca3a086010000000000225120b23da6d2e0390018b953f7d74e3582da4da30fd0fd157cc84a2d2753003d1ca3a086010000000000225120b23da6d2e0390018b953f7d74e3582da4da30fd0fd157cc84a2d2753003d1ca3a086010000000000225120b23da6d2e0390018b953f7d74e3582da4da30fd0fd157cc84a2d2753003d1ca3a086010000000000225120b23da6d2e0390018b953f7d74e3582da4da30fd0fd157cc84a2d2753003d1ca35c081777000000002251202a64b1ee3375f3bb4b367b8cb8384a47f73cf231717f827c6c6fbbf5aecf0c364a010000000000002200204ae81572f06e1b88fd5ced7a1a000945432e83e1551e6f721ee9c00b8cc33260014005a41e6f4a4bcfcc5cd3ef602687215f97c18949019a491df56af7413c5dce9292ba3966edc4564a39d9bc0d6c0faae19030f1cedf4d931a6cdc57cc5b83c8ef00000000".to_string();
@@ -949,7 +946,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_deposit_kickoff_generator_tx_1() {
-        let config = create_test_config_with_thread_name!(None);
+        let config = create_test_config_with_thread_name(None).await;
         let db = Database::new(&config).await.unwrap();
 
         let raw_hex = "01000000000101308d840c736eefd114a8fad04cb0d8338b4a3034a2b517250e5498701b25eb360100000000fdffffff02401f00000000000022512024985a1ab5724a5164ae5e0026b3e7e22031e83948eedf99d438b866857946b81f7e000000000000225120f7298da2a2be5b6e02a076ff7d35a1fe6b54a2bc7938c1c86bede23cadb7d9650140ad2fdb01ec5e2772f682867c8c6f30697c63f622e338f7390d3abc6c905b9fd7e96496fdc34cb9e872387758a6a334ec1307b3505b73121e0264fe2ba546d78ad11b0d00".to_string();
@@ -999,7 +996,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_operators_kickoff_utxo_1() {
-        let config = create_test_config_with_thread_name!(None);
+        let config = create_test_config_with_thread_name(None).await;
         let db = Database::new(&config).await.unwrap();
 
         let outpoint = OutPoint {
@@ -1024,7 +1021,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_operators_kickoff_utxo_2() {
-        let config = create_test_config_with_thread_name!(None);
+        let config = create_test_config_with_thread_name(None).await;
         let db = Database::new(&config).await.unwrap();
 
         let outpoint = OutPoint {
@@ -1037,7 +1034,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_operators_funding_utxo_1() {
-        let config = create_test_config_with_thread_name!(None);
+        let config = create_test_config_with_thread_name(None).await;
         let db = Database::new(&config).await.unwrap();
 
         let utxo = UTXO {
@@ -1059,7 +1056,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_operators_funding_utxo_2() {
-        let config = create_test_config_with_thread_name!(None);
+        let config = create_test_config_with_thread_name(None).await;
         let db = Database::new(&config).await.unwrap();
 
         let db_utxo = db.get_funding_utxo(None).await.unwrap();
@@ -1069,7 +1066,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_deposit_kickoff_generator_tx_2() {
-        let config = create_test_config_with_thread_name!(None);
+        let config = create_test_config_with_thread_name(None).await;
         let db = Database::new(&config).await.unwrap();
 
         let txid = Txid::from_byte_array([1u8; 32]);
@@ -1079,7 +1076,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_deposit_kickoff_generator_tx_3() {
-        let config = create_test_config_with_thread_name!(None);
+        let config = create_test_config_with_thread_name(None).await;
         let db = Database::new(&config).await.unwrap();
 
         let txid = Txid::from_byte_array([1u8; 32]);
@@ -1089,7 +1086,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_save_get_bitvm_setup() {
-        let config = create_test_config_with_thread_name!(None);
+        let config = create_test_config_with_thread_name(None).await;
         let database = Database::new(&config).await.unwrap();
 
         let operator_idx = 0;
@@ -1150,9 +1147,9 @@ mod tests {
 
     #[tokio::test]
     async fn set_get_operator_winternitz_public_keys() {
-        let mut config = create_test_config_with_thread_name!(None);
+        let mut config = create_test_config_with_thread_name(None).await;
         let database = Database::new(&config).await.unwrap();
-        let regtest = create_regtest_rpc!(config);
+        let regtest = create_regtest_rpc(&mut config).await;
         let rpc = regtest.rpc().clone();
 
         let operator = Operator::new(config, rpc).await.unwrap();
@@ -1180,7 +1177,7 @@ mod tests {
 
     #[tokio::test]
     async fn set_get_deposit_signatures() {
-        let config = create_test_config_with_thread_name!(None);
+        let config = create_test_config_with_thread_name(None).await;
         let database = Database::new(&config).await.unwrap();
 
         let operator_idx = 0x45;
