+++ core/src/database/verifier.rs
@@ -273,11 +273,8 @@ impl Database {
 
 #[cfg(test)]
 mod tests {
-    use crate::{config::BridgeConfig, initialize_database, utils::initialize_logger};
-    use crate::{
-        create_test_config_with_thread_name, database::Database, musig2::nonce_pair, utils::SECP,
-        EVMAddress, UTXO,
-    };
+    use crate::test::common::*;
+    use crate::{database::Database, musig2::nonce_pair, utils::SECP, EVMAddress, UTXO};
     use bitcoin::{
         hashes::Hash, key::Keypair, secp256k1::SecretKey, Address, Amount, OutPoint, ScriptBuf,
         TxOut, Txid, XOnlyPublicKey,
@@ -287,7 +284,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_verifiers_kickoff_utxos_1() {
-        let config = create_test_config_with_thread_name!(None);
+        let config = create_test_config_with_thread_name(None).await;
         let db = Database::new(&config).await.unwrap();
 
         let outpoint = OutPoint {
@@ -324,7 +321,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_verifiers_kickoff_utxos_2() {
-        let config = create_test_config_with_thread_name!(None);
+        let config = create_test_config_with_thread_name(None).await;
         let db = Database::new(&config).await.unwrap();
 
         let outpoint = OutPoint {
@@ -337,7 +334,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_get_pub_nonces_1() {
-        let config = create_test_config_with_thread_name!(None);
+        let config = create_test_config_with_thread_name(None).await;
         let db = Database::new(&config).await.unwrap();
 
         let outpoint = OutPoint {
@@ -369,7 +366,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_get_pub_nonces_2() {
-        let config = create_test_config_with_thread_name!(None);
+        let config = create_test_config_with_thread_name(None).await;
         let db = Database::new(&config).await.unwrap();
         let outpoint = OutPoint {
             txid: Txid::from_byte_array([1u8; 32]),
@@ -381,7 +378,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_save_and_get_deposit_info() {
-        let config = create_test_config_with_thread_name!(None);
+        let config = create_test_config_with_thread_name(None).await;
         let database = Database::new(&config).await.unwrap();
 
         let xonly_public_key = XOnlyPublicKey::from_slice(&[
