+++ core/src/test/mod.rs
@@ -1,0 +1,6 @@
+pub mod common;
+mod deposit;
+mod misc;
+mod musig2;
+mod rpc;
+mod taproot;
