+++ core/src/test/musig2.rs
@@ -1,30 +1,27 @@
+use crate::builder::script::{CheckSig, OtherSpendable, SpendPath, SpendableScript};
+use crate::builder::transaction::input::SpendableTxIn;
+use crate::builder::transaction::output::UnspentTxOut;
+use crate::builder::transaction::{TransactionType, TxHandlerBuilder, DEFAULT_SEQUENCE};
+use crate::errors::BridgeError;
+use crate::musig2::{
+    aggregate_nonces, aggregate_partial_signatures, AggregateFromPublicKeys, Musig2Mode,
+};
+use crate::rpc::clementine::NormalSignatureKind;
+use crate::test::common::*;
+use crate::utils::SECP;
+use crate::{
+    builder::{self},
+    config::BridgeConfig,
+    musig2::{nonce_pair, partial_sign, MuSigNoncePair},
+    utils,
+};
 use bitcoin::key::Keypair;
 use bitcoin::secp256k1::{Message, PublicKey};
 use bitcoin::{hashes::Hash, script, Amount, TapSighashType};
 use bitcoin::{taproot, Sequence, TxOut, XOnlyPublicKey};
 use bitcoincore_rpc::RpcApi;
-use clementine_core::builder::script::{CheckSig, OtherSpendable, SpendPath, SpendableScript};
-use clementine_core::builder::transaction::input::SpendableTxIn;
-use clementine_core::builder::transaction::output::UnspentTxOut;
-use clementine_core::builder::transaction::{TransactionType, TxHandlerBuilder, DEFAULT_SEQUENCE};
-use clementine_core::errors::BridgeError;
-use clementine_core::musig2::{
-    aggregate_nonces, aggregate_partial_signatures, AggregateFromPublicKeys, Musig2Mode,
-};
-use clementine_core::rpc::clementine::NormalSignatureKind;
-use clementine_core::utils::SECP;
-use clementine_core::{
-    builder::{self},
-    config::BridgeConfig,
-    extended_rpc::ExtendedRpc,
-    musig2::{nonce_pair, partial_sign, MuSigNoncePair},
-    utils,
-};
-use clementine_core::{database::Database, utils::initialize_logger};
 use secp256k1::musig::{MusigAggNonce, MusigPartialSignature};
 use std::sync::Arc;
-
-mod common;
 
 #[cfg(test)]
 fn get_verifiers_keys(config: &BridgeConfig) -> (Vec<Keypair>, XOnlyPublicKey, Vec<PublicKey>) {
@@ -71,10 +68,9 @@ fn get_nonces(
 }
 
 #[tokio::test]
-
 async fn key_spend() {
-    let mut config = create_test_config_with_thread_name!(None);
-    let regtest = create_regtest_rpc!(config);
+    let mut config = create_test_config_with_thread_name(None).await;
+    let regtest = create_regtest_rpc(&mut config).await;
     let rpc = regtest.rpc().clone();
 
     let (verifiers_secret_public_keys, untweaked_xonly_pubkey, verifier_public_keys) =
@@ -168,8 +164,8 @@ async fn key_spend() {
 #[tokio::test]
 
 async fn key_spend_with_script() {
-    let mut config = create_test_config_with_thread_name!(None);
-    let regtest = create_regtest_rpc!(config);
+    let mut config = create_test_config_with_thread_name(None).await;
+    let regtest = create_regtest_rpc(&mut config).await;
     let rpc = regtest.rpc().clone();
 
     let (verifiers_secret_public_keys, untweaked_xonly_pubkey, verifier_public_keys) =
@@ -273,8 +269,8 @@ async fn key_spend_with_script() {
 #[tokio::test]
 
 async fn script_spend() {
-    let mut config = create_test_config_with_thread_name!(None);
-    let regtest = create_regtest_rpc!(config);
+    let mut config = create_test_config_with_thread_name(None).await;
+    let regtest = create_regtest_rpc(&mut config).await;
     let rpc = regtest.rpc().clone();
 
     let (verifiers_secret_public_keys, _untweaked_xonly_pubkey, verifier_public_keys) =
@@ -382,8 +378,8 @@ async fn key_and_script_spend() {
     use bitcoin::{Network::*, *};
 
     // Arrange
-    let mut config = create_test_config_with_thread_name!(None);
-    let regtest = create_regtest_rpc!(config);
+    let mut config = create_test_config_with_thread_name(None).await;
+    let regtest = create_regtest_rpc(&mut config).await;
     let rpc = regtest.rpc().clone();
 
     // -- Musig2 Setup --
