+++ core/src/tx_sender.rs
@@ -687,12 +687,8 @@ mod tests {
     use crate::bitcoin_syncer;
     use crate::builder::script::{CheckSig, SpendableScript};
     use crate::builder::transaction::TransactionType;
-    use crate::config::BridgeConfig;
-    use crate::utils::{initialize_logger, SECP};
-    use crate::{
-        create_regtest_rpc, create_test_config_with_thread_name, database::Database,
-        initialize_database,
-    };
+    use crate::utils::SECP;
+    use crate::{database::Database, test::common::*};
     use bitcoin::secp256k1::SecretKey;
     use bitcoin::transaction::Version;
     use secp256k1::rand;
@@ -705,7 +701,7 @@ mod tests {
         let network = bitcoin::Network::Regtest;
         let actor = Actor::new(sk, None, network);
 
-        let config = create_test_config_with_thread_name!(None);
+        let config = create_test_config_with_thread_name(None).await;
 
         let db = Database::new(&config).await.unwrap();
 
@@ -778,8 +774,8 @@ mod tests {
 
     #[tokio::test]
     async fn test_try_to_send() -> Result<(), BridgeError> {
-        let mut config = create_test_config_with_thread_name!(None);
-        let regtest = create_regtest_rpc!(config);
+        let mut config = create_test_config_with_thread_name(None).await;
+        let regtest = create_regtest_rpc(&mut config).await;
         let rpc = regtest.rpc().clone();
 
         rpc.mine_blocks(1).await.unwrap();
@@ -866,8 +862,8 @@ mod tests {
 
     #[tokio::test]
     async fn get_fee_rate() {
-        let mut config = create_test_config_with_thread_name!(None);
-        let regtest = create_regtest_rpc!(config);
+        let mut config = create_test_config_with_thread_name(None).await;
+        let regtest = create_regtest_rpc(&mut config).await;
         let rpc: ExtendedRpc = regtest.rpc().clone();
         let db = Database::new(&config).await.unwrap();
