+++ core/src/verifier.rs
@@ -1409,14 +1409,14 @@ impl Verifier {
 // use crate::EVMAddress;
 // use crate::{actor::Actor, create_test_config_with_thread_name};
 // use crate::{
-//     config::BridgeConfig, database::Database, initialize_database, utils::initialize_logger,
+//     config::BridgeConfig, database::Database, test::common::*, utils::initialize_logger,
 // };
 // use bitcoin::secp256k1:: rand;
 // use std::{env, thread};
 
 // #[tokio::test]
 // async fn verifier_new_public_key_check() {
-//     let mut config = create_test_config_with_thread_name!(None);
+//     let mut config = create_test_config_with_thread_name(None).await;
 //     let rpc = ExtendedRpc::connect(
 //         config.bitcoin_rpc_url.clone(),
 //         config.bitcoin_rpc_user.clone(),
@@ -1435,7 +1435,7 @@ impl Verifier {
 // #[tokio::test]
 //
 // async fn new_deposit_nonce_checks() {
-//     let mut config = create_test_config_with_thread_name!(None);
+//     let mut config = create_test_config_with_thread_name(None).await;
 //     let rpc = ExtendedRpc::connect(
 //         config.bitcoin_rpc_url.clone(),
 //         config.bitcoin_rpc_user.clone(),
@@ -1445,7 +1445,7 @@ impl Verifier {
 //     let verifier = Verifier::new(rpc.clone_inner().await.unwrap(), config.clone()).await.unwrap();
 
 //     let evm_address = EVMAddress([1u8; 20]);
-//     let deposit_address = get_deposit_address(config, evm_address).unwrap(); This line needs to be converted into get_deposit_address!
+//     let deposit_address = get_deposit_address(config, evm_address).unwrap(); This line needs to be converted into get_deposit_address
 
 //     let signer_address = Actor::new(
 //         config.secret_key,
